moving the record down, so that the triple is not in any interfaces
