+++ graylog2-server/src/main/java/org/graylog/security/certutil/CertRenewalService.java
@@ -22,10 +22,13 @@ import org.graylog2.cluster.Node;
 import org.graylog2.cluster.preflight.DataNodeProvisioningConfig;
 
 import java.security.cert.X509Certificate;
+import java.time.LocalDateTime;
 import java.util.List;
 
 public interface CertRenewalService {
+    record DataNode(String nodeId, Node.Type type, String transportAddress, DataNodeProvisioningConfig.State status, String errorMsg, String hostname, String shortNodeId, LocalDateTime certValidUntil) {}
+
     void checkCertificatesForRenewal();
     void initiateRenewalForNode(String nodeId);
-    List<Triple<Node, DataNodeProvisioningConfig, X509Certificate>> findNodes();
+    List<DataNode> findNodes();
 }
